try to fix codeQL
