+++ shared/src/main/java/cn/edu/tsinghua/iginx/utils/FileUtils.java
@@ -4,7 +4,6 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.file.*;
-import java.nio.file.attribute.BasicFileAttributes;
 import java.util.List;
 
 public class FileUtils {
@@ -51,33 +50,6 @@ public class FileUtils {
     }
   }
 
-  public static void copyFileOrDir(File source, File target) throws IOException {
-    if (source.isFile() && target.isFile()) {
-      Files.copy(source.toPath(), target.toPath());
-      return;
-    }
-    Files.walkFileTree(
-        source.toPath(),
-        new SimpleFileVisitor<Path>() {
-          @Override
-          public FileVisitResult preVisitDirectory(final Path dir, final BasicFileAttributes attrs)
-              throws IOException {
-            Files.createDirectories(target.toPath().resolve(source.toPath().relativize(dir)));
-            return FileVisitResult.CONTINUE;
-          }
-
-          @Override
-          public FileVisitResult visitFile(final Path file, final BasicFileAttributes attrs)
-              throws IOException {
-            Files.copy(
-                file,
-                target.toPath().resolve(source.toPath().relativize(file)),
-                StandardCopyOption.REPLACE_EXISTING);
-            return FileVisitResult.CONTINUE;
-          }
-        });
-  }
-
   public static void deleteFileOrDir(File file) throws IOException {
     if (file.isFile()) {
       if (!file.delete()) {
@@ -89,20 +61,6 @@ public class FileUtils {
   }
 
   public static void deleteFolder(File folder) throws IOException {
-    File[] files = folder.listFiles();
-    if (files != null) {
-      for (File f : files) {
-        if (f.isDirectory()) {
-          deleteFolder(f);
-        } else {
-          if (!f.delete()) {
-            throw new IOException("Failed to delete file: " + f);
-          }
-        }
-      }
-    }
-    if (!folder.delete()) {
-      throw new IOException("Failed to delete the folder: " + folder);
-    }
+    org.apache.commons.io.FileUtils.deleteDirectory(folder);
   }
 }
